Fix the delete() operation
After the current record is deleted, it's not acceptable to perform any
operations on it any more (i.e. delete() followed by set_rr_ttl() doesn't
make any sense).

Ideally, the compiler should refuse code with operations sitting between
delete() and the end of the loop block.

Right now, we intentionally panic when it happens (because the offset is
reset to None). For safety, we should either return a soft error, or
make all operations no-ops.