Fix a pesky bug when insertion when rr_len == insertion offset
I got confused by the Vec::reserve() API that doesn't take the new
length, but the additional length.

As a result, inserting synthetic records always worked except under
very specific conditions, that didn't necessarily lead to invalid
packets. Especially when an OPT section was present.

That bug was really painful to nail down, but the fix is trivial.

Damn it.