Expose something more convenient than iter_question in the C API
Because there will always be at most 1 question, iterating over the
section would be a waste of CPU cycles and makes thing more
complicated than they should to users.

We will expose a question() C API that returns the question, with
the name as a string, and the query type.

Easier, better, faster.